4.21

+++ part4/blog/controllers/blog.js
@@ -39,10 +39,22 @@ blogRouter.post('/', async (request, response, next) => {
 
 blogRouter.delete('/:id', async (request, response, next) => {
   const blogToDelete = request.params.id;
-  const result = await Blog.deleteOne({
-    _id: blogToDelete
-  });
-  return response.status(204).end();
+  const token = request.token;
+  const decodedToken = jwt.verify(token, process.env.SECRET);
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+
+  const userToken = await User.findById(decodedToken.id);
+  const blog = await Blog.findById(blogToDelete);
+
+  if (blog.user.toString() === userToken.id.toString()) {
+   await Blog.deleteOne({
+      _id: blogToDelete
+    });
+   return response.status(204).end();
+  }
+  return response.status(401).json({ error: 'invalid user'});
 });
 
 blogRouter.put('/:id', async (request, response, next) => {
